Added check for non graded assignmnet

+++ common/src/main/java/com/scholarscore/models/IGoal.java
@@ -11,7 +11,7 @@ import javax.persistence.*;
 public interface IGoal {
 
     @Transient
-    public Long getCalculatedValue();
+    public Float getCalculatedValue();
